requirements.txt
discord.py
watchfiles
PyYAML
aiofiles
b2sdk
aiohttp